+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-08-11 15:34:33
-// Commit: style(state): harmonize light and dark theme palette for state
+// Updated at: 2026-08-11 19:04:51
+// Commit: chore(state): update dependencies and cleanup unused imports in state
 
 class stateComponentState {
-  final String commitRef = '3532';
-  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
-  final DateTime timestamp = DateTime.parse('2026-08-11T15:34:33+05:00');
+  final String commitRef = '3543';
+  final String lastMessage = 'chore(state): update dependencies and cleanup unused imports in state';
+  final DateTime timestamp = DateTime.parse('2026-08-11T19:04:51+05:00');
 }